Split core/actions/table_test.ts and core/actions/incremental_table_test.ts
