+++ core/BUILD
@@ -70,11 +70,15 @@ ts_test_suite(
         "actions/data_preparation_test.ts",
         "actions/declaration_test.ts",
         "actions/filename_override_test.ts",
+        "actions/incremental_table_config_test.ts",
+        "actions/incremental_table_iceberg_test.ts",
         "actions/incremental_table_test.ts",
         "actions/index_test.ts",
         "actions/notebook_test.ts",
         "actions/operation_test.ts",
         "actions/property_graph_test.ts",
+        "actions/table_config_test.ts",
+        "actions/table_iceberg_test.ts",
         "actions/table_test.ts",
         "actions/test_test.ts",
         "actions/view_test.ts",
